reorganize directory structure
So, this was originally structured (and maybe it's the Django way,
since it was so easy to stumble into this configuration with the
standard `startproject` command??) so that the root of the repository
contained manage.py, my virtualenv stuff, and a Finetooth/ directory
where the actual code lives. But I feel like that's unnecessary
nesting: I want the root of the repository to be useful; having
manage.py and settings.py in the same directory makes sense to me. Of
course, moving stuff required changing a lot of module import paths,
too: I think I got it all, because I can retrieve models at the
manage.py shell and the site shows the posts again (and this project
doesn't really do anything else at this point).

(Oh, and I also just tested that I can make and apply a migration #2.)

Maybe the extra layer of nesting would have been useful for dependency
management? If my code is going to live in a "core" app, then
importing some other Python module that happens to be named "core"
would probably require some weird import logic hacking? But I can't
quite bring myself to care.