starting to port color functions; tests directory
Later, we'll want to interpolate between colors and generate CSS in
order to display substrings with different scores differently. I've
already written code that solves this problem (for a different
project, in an a sort-of different language; see
https://github.com/zackmdavis/Thirty-Three/blob/439c490e/buildscript.hy),
so I'll just translate/reuse it.

Also, let's have a tests directory instead of a single file. I don't
like the test discovery behavior that's apparently new in Django 1.6
(where test files are assumed to start with "test"), but it's easy
enough to work around.