fix indentation to four spaces (!?!)
I'm kind of shocked that the interpreter let this happen at all (and
disappointed in Emacs, too). It looks like the interpreter actually
just tracks relative indentation (the "How does the compiler parse the
indentation?" section in
http://www.secnetix.de/olli/Python/block_indentation.hawk was
instructive on this point) rather than enforcing consistently
"globally", which totally makes sense in retrospect!