towards voting feedback
The JavaScript library directory is called libs/ instead of lib/
because of my personal beliefs about how .gitignore works and how I
had wanted to put the borrowed snippet that changes the Underscore
templating delimiters in its own file there in order to avoid
duplication, because Jasmine needs to know about it, too, not just
base.html. But then I actually couldn't figure out how to get Jasmine
to recognize it (you'd think (apparently having an episodic memory of)
putting it in the src_files list in jasmine.yml would be enough!?), so
it's at the top of finetooth.js for now.