first cut at pagination
In the template, I had actually hoped to use the {%
bootstrap_pagination %} template tags provided by the
django-bootstrap3 app that we're using, but the documentation was
uninformative on how to use it, and glancing at the source, it kind of
seems like it assumes that pages are given in the query string (am I
misunderstanding something?!), so I guess we'll just have to do it
ourselves. What we have in this commit isn't quite right (for example,
there's a link to page zero, which 500s on an EmptyPage "That page
number is less than 1" exception), but it's an okayish start. We'll
also want to abstract this into a function (as with `scored_context`).

[#21]