:hourglass: color tags by score, too (very slowly)
This commit hacks together a way for tags to be displayed in color
based on the scores of associated posts. Absolutely no regard was
given to efficiency, however (the IndirectlyScorableMixin.{high,
low}_score methods are _especially_ ridiculous), which might explain
why the homepage takes ten seconds to load (contrast to 0.4 seconds on
master). A great opportunity to lean about profiling??

[#35]