update README; forget sample data; clean commented-out junk
7fead4b7 was due to the sample data having been rotted by more recent
model changes; having a sample data fixture isn't cool enough to
bother with keeping it up to date (and we have factories now anyway,
which are much better for developers who want development data and
could in any case be used to serve the purpose of created a
WordPress-inspired "Welcome to Finetooth" post, which was the original
idea), so let's just not bother.

I'm ashamed about commiting (and pushing to master! and tagging!) that
commented-out code snippet, but if it helps, I was pretty tired.