hasty but obligatory Django 2.0 upgrade
The future is now!

Squash-regenerating as one initial migration was easier than figuring out how
to edit the migration history now that User.on_delete is mandatory (recall
4d6a2a8a); I don't know what people do in production.

It doesn't look like the debug toolbar is working, and we should feel guilty
for using a deprecation mixin ... but let it pass for now.

Resolves #65.